+++ src/Telegram.Bot/Types/ReplyMarkups/Buttons/KeyboardButton.cs
@@ -14,23 +14,6 @@ namespace Telegram.Bot.Types.ReplyMarkups.Buttons
         public string Text { get; set; }
 
         /// <summary>
-        /// Performs an implicit conversion from <see cref="string"/> to <see cref="KeyboardButton"/>.
-        /// </summary>
-        /// <param name="key">The key.</param>
-        /// <returns>
-        /// The result of the conversion.
-        /// </returns>
-        public static implicit operator KeyboardButton(string key)
-            => new KeyboardButton(key);
-
-        /// <summary>
-        /// Performs an implicit conversion from <see cref="InlineKeyboardButton"/> to <see cref="KeyboardButton"/>.
-        /// </summary>
-        /// <param name="button"></param>
-        public static implicit operator KeyboardButton(InlineKeyboardButton button)
-            => new KeyboardButton(button.Text);
-
-        /// <summary>
         /// Initializes a new instance of the <see cref="KeyboardButton"/> class.
         /// </summary>
         /// <param name="text">Label text on the button</param>
@@ -38,5 +21,15 @@ namespace Telegram.Bot.Types.ReplyMarkups.Buttons
         {
             Text = text;
         }
+
+        /// <summary>
+        /// Performs an implicit conversion from <see cref="string"/> to <see cref="KeyboardButton"/>.
+        /// </summary>
+        /// <param name="text">The text.</param>
+        /// <returns>
+        /// The result of the conversion.
+        /// </returns>
+        public static implicit operator KeyboardButton(string text)
+            => new KeyboardButton(text);
     }
 }