+++ src/Telegram.Bot/Types/ReplyMarkups/ForceReplyMarkup.cs
@@ -6,14 +6,13 @@ namespace Telegram.Bot.Types.ReplyMarkups
     /// <summary>
     /// Upon receiving a <see cref="Message"/> with this object, Telegram clients will display a reply interface to the user (act as if the user has selected the bot's message and tapped 'Reply'). This can be extremely useful if you want to create user-friendly step-by-step interfaces without having to sacrifice privacy mode.
     /// </summary>
-    [JsonObject(MemberSerialization = MemberSerialization.OptIn,
-                NamingStrategyType = typeof(SnakeCaseNamingStrategy))]
-    public class ForceReply : ReplyMarkup
+    [JsonObject(MemberSerialization.OptIn, NamingStrategyType = typeof(SnakeCaseNamingStrategy))]
+    public class ForceReplyMarkup : ReplyMarkupBase
     {
         /// <summary>
         /// Shows reply interface to the user, as if they manually selected the bot's message and tapped 'Reply'
         /// </summary>
-        [JsonProperty(PropertyName = "force_reply", Required = Required.Always)]
-        public bool Force { get; set; }
+        [JsonProperty(Required = Required.Always)]
+        public bool ForceReply => true;
     }
 }