+++ src/Telegram.Bot/Types/ReplyMarkups/IReplyMarkup.cs
@@ -1,7 +1,7 @@
 ﻿namespace Telegram.Bot.Types.ReplyMarkups
 {
     /// <summary>
-    /// Objects implementing this Interface define how a <see cref="User"/> can reply to the sent <see cref="Message"/>
+    /// A marker interface for reply markups that define how a <see cref="User"/> can reply to the sent <see cref="Message"/>
     /// </summary>
     public interface IReplyMarkup
     {